chore: separate auth check in it's own file
